vps run

+++ neuralnetwork/nn_modelling.py
@@ -68,8 +68,8 @@ eras_to_embargo = [str(era).zfill(4) for era in [last_train_era + i for i in ran
 validation = validation[~validation["era"].isin(eras_to_embargo)]
 
 
-X_test = torch.tensor(validation[feature_cols].values)
-y_test = torch.tensor(validation[target].values).reshape(-1, 1)
+X_test = torch.tensor(validation[feature_cols].values, dtype = torch.float32)
+y_test = torch.tensor(validation[target].values, dtype = torch.float32).reshape(-1, 1)
 
 #X_test = torch.tensor(X_test.values, dtype=torch.float32)
 #y_test = torch.tensor(y_test.values, dtype=torch.float32).reshape(-1, 1)
